chore: separate connection handler
