+++ src/module/consumer.ts
@@ -3,6 +3,7 @@ import { getLogger } from '@/util/logger';
 import { handler as historyBatchProcesser } from '@/handlers/history-batch-processer';
 import { getPgPool } from '@/lib/pg';
 import { ConsumeMessage } from 'amqplib';
+import Amqp from '@/lib/amqp';
 
 const logger = getLogger('consumer');
 const pgPool = getPgPool();
@@ -16,12 +17,16 @@ const PREFETCH_COUNT = 20;
 const BATCH_SIZE = 10;
 const BATCH_TIMEOUT_MS = 10000;
 
-let batchConsumer: BatchConsumer | null = null;
+let connection: Amqp | null = null;
 
 export async function startConsumer(): Promise<void> {
   if (!pgPool) {
     throw new Error('Pg pool not initialized');
   }
+
+  connection = await Amqp.createClient(RABBIT_MQ_CONNECTION_STRING);
+
+  const channel = connection.getChannel();
 
   const batchConsumerOptions: BatchConsumerOptions = {
     amqpConnectionString: RABBIT_MQ_CONNECTION_STRING,
@@ -36,19 +41,23 @@ export async function startConsumer(): Promise<void> {
     batchTimeoutMs: BATCH_TIMEOUT_MS
   };
 
-  batchConsumer = new BatchConsumer(batchConsumerOptions, (messages: ConsumeMessage[]) =>
-    historyBatchProcesser(pgPool, messages)
+  const batchConsumer = new BatchConsumer(
+    channel,
+    batchConsumerOptions,
+    (messages: ConsumeMessage[]) => historyBatchProcesser(pgPool, messages)
   );
+
+  await batchConsumer.start();
 }
 
 export async function stopConsumer(): Promise<void> {
-  if (batchConsumer) {
+  if (connection) {
     try {
-      await batchConsumer.close();
+      await connection.close();
     } catch (err) {
       logger.error('Error closing batch consumer:', { err });
     } finally {
-      batchConsumer = null;
+      connection = null;
     }
   }
 }
